Phase 01-Test
Task: Build a Form and its validation. We don’t need to save the data. On submission the form will be validated and if there are any errors, there will be an alert that tells the user what the issue in the form is and how they can resolve that issue.